+++ http.go
@@ -2,12 +2,10 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"html/template"
 	"log"
 	"net/http"
 	"os"
-	"strconv"
 	"time"
 
 	"github.com/shurcooL/httpfs/html/vfstemplate"
@@ -167,350 +165,5 @@ func webserver() {
 		if err != nil {
 			log.Fatal(err)
 		}
-	}
-}
-
-// handleSearch는 URL을 통해 query를 할 수 있게 해주는 함수입니다.
-func handleSearch(w http.ResponseWriter, r *http.Request) {
-	token, err := GetTokenFromHeader(w, r)
-	if err != nil {
-		http.Redirect(w, r, "/signin", http.StatusSeeOther)
-		return
-	}
-	q := r.URL.Query()
-	itemType := q.Get("itemtype")
-	searchword := q.Get("searchword")
-	page := PageToString(q.Get("page"))
-	if itemType == "" {
-		itemType = "maya"
-	}
-	if page == "" {
-		page = "1"
-	}
-	type recipe struct {
-		Items       []Item
-		Searchword  string
-		ItemType    string
-		TotalNum    int64
-		CurrentPage int64
-		TotalPage   int64
-		Pages       []int64
-		Token
-		User         User
-		Adminsetting Adminsetting
-	}
-	rcp := recipe{}
-	rcp.Searchword = searchword
-	rcp.ItemType = itemType
-	rcp.Token = token
-	//mongoDB client 연결
-	client, err := mongo.NewClient(options.Client().ApplyURI(*flagMongoDBURI))
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	err = client.Connect(ctx)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	defer client.Disconnect(ctx)
-	err = client.Ping(ctx, readpref.Primary())
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	rcp.CurrentPage = PageToInt(page)
-	totalPage, totalNum, items, err := SearchPage(client, itemType, searchword, rcp.CurrentPage, *flagPagenum)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	rcp.Items = items
-	rcp.User, err = GetUser(client, token.ID) // user 정보 가져옴
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-	}
-	rcp.TotalNum = totalNum
-	rcp.TotalPage = totalPage
-	// Pages를 설정한다.
-	rcp.Pages = make([]int64, totalPage) // page에 필요한 메모리를 미리 설정한다.
-	for i := range rcp.Pages {
-		rcp.Pages[i] = int64(i) + 1
-	}
-	adminsetting, err := GetAdminSetting(client)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	rcp.Adminsetting = adminsetting
-	err = TEMPLATES.ExecuteTemplate(w, "index", rcp)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func handleSearchSubmit(w http.ResponseWriter, r *http.Request) {
-	_, err := GetTokenFromHeader(w, r)
-	if err != nil {
-		http.Redirect(w, r, "/signin", http.StatusSeeOther)
-		return
-	}
-	itemType := r.FormValue("itemtype")
-	searchword := r.FormValue("searchword")
-	page := PageToString(r.FormValue("page"))
-	http.Redirect(w, r, fmt.Sprintf("/search?itemtype=%s&searchword=%s&page=%s", itemType, searchword, page), http.StatusSeeOther)
-}
-
-func handleHelp(w http.ResponseWriter, r *http.Request) {
-	token, err := GetTokenFromHeader(w, r)
-	if err != nil {
-		http.Redirect(w, r, "/signin", http.StatusSeeOther)
-		return
-	}
-	type recipe struct {
-		Token
-		Adminsetting Adminsetting
-	}
-	rcp := recipe{}
-	rcp.Token = token
-	//mongoDB client 연결
-	client, err := mongo.NewClient(options.Client().ApplyURI(*flagMongoDBURI))
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	err = client.Connect(ctx)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	defer client.Disconnect(ctx)
-	err = client.Ping(ctx, readpref.Primary())
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	adminsetting, err := GetAdminSetting(client)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	rcp.Adminsetting = adminsetting
-	w.Header().Set("Content-Type", "text/html")
-	err = TEMPLATES.ExecuteTemplate(w, "help", rcp)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func handleItemProcess(w http.ResponseWriter, r *http.Request) {
-	token, err := GetTokenFromHeader(w, r)
-	if err != nil {
-		http.Redirect(w, r, "/signin", http.StatusSeeOther)
-		return
-	}
-	w.Header().Set("Content-Type", "text/html")
-	type recipe struct {
-		Items []Item
-		Token
-		Adminsetting   Adminsetting
-		StorageTitle   string
-		StoragePercent int64
-		StorageLevel   int64
-	}
-	//mongoDB client 연결
-	client, err := mongo.NewClient(options.Client().ApplyURI(*flagMongoDBURI))
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	err = client.Connect(ctx)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	defer client.Disconnect(ctx)
-	err = client.Ping(ctx, readpref.Primary())
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	rcp := recipe{}
-	// 완료되지 않은 아이템을 가져온다
-	rcp.Items, err = GetOngoingProcess(client)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	rcp.Token = token
-	adminsetting, err := GetAdminSetting(client)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	rcp.Adminsetting = adminsetting
-	ds, err := DiskCheck()
-	if err != nil {
-		rcp.StorageTitle = "Storage Usage (Please set RootPath)"
-		rcp.StoragePercent = 0
-		rcp.StorageLevel = 0
-	} else {
-		rcp.StorageTitle = "Storage Usage"
-		rcp.StoragePercent = int64((float64(ds.Used) / float64(ds.All)) * 100)
-		rcp.StorageLevel = rcp.StoragePercent / 10
-	}
-	err = TEMPLATES.ExecuteTemplate(w, "item-process", rcp)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func handleCleanUpDB(w http.ResponseWriter, r *http.Request) {
-	token, err := GetTokenFromHeader(w, r)
-	if err != nil {
-		http.Redirect(w, r, "/signin", http.StatusSeeOther)
-		return
-	}
-	w.Header().Set("Content-Type", "text/html")
-	type recipe struct {
-		Items []Item
-		Token
-		Adminsetting Adminsetting
-	}
-	//mongoDB client 연결
-	client, err := mongo.NewClient(options.Client().ApplyURI(*flagMongoDBURI))
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	err = client.Connect(ctx)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	defer client.Disconnect(ctx)
-	rcp := recipe{}
-	// 데이터가 모두 업로드되지 않은 아이템을 가져온다
-	rcp.Items, err = GetIncompleteItems(client)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	rcp.Token = token
-	adminsetting, err := GetAdminSetting(client)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	rcp.Adminsetting = adminsetting
-	err = TEMPLATES.ExecuteTemplate(w, "cleanup-db", rcp)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func handleCleanUpDBSubmit(w http.ResponseWriter, r *http.Request) {
-	_, err := GetTokenFromHeader(w, r)
-	if err != nil {
-		http.Redirect(w, r, "/signin", http.StatusSeeOther)
-		return
-	}
-	//checkbox의 개수를 받아온다.
-	itemNum, err := strconv.Atoi(r.FormValue("itemnum"))
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	itemsToDelete := make(map[string]string)
-	//checkbox의 개수만큼 포문을 돈다
-	for i := 0; i < itemNum; i++ {
-		//check되어 있는지 확인한다.
-		id := r.FormValue(fmt.Sprintf("checkbox%d", i+1))
-		if id == "" {
-			continue // check 안되어 있으면 넘어간다
-		} //check 되어 있으면 itemytpe을 가져와서 map을 채운다.
-		itemtype := r.FormValue(fmt.Sprintf("itemtype%d", i+1))
-		itemsToDelete[id] = itemtype
-	}
-	//mongoDB client 연결
-	client, err := mongo.NewClient(options.Client().ApplyURI(*flagMongoDBURI))
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	err = client.Connect(ctx)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	defer client.Disconnect(ctx)
-	//삭제할 아이템의 리스트를 돌면서 삭제한다.
-	for id, itemtype := range itemsToDelete {
-		err = RmItem(client, itemtype, id)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-	}
-	http.Redirect(w, r, "/cleanup-db", http.StatusSeeOther)
-}
-
-func handleCleanUpDBSuccess(w http.ResponseWriter, r *http.Request) {
-	token, err := GetTokenFromHeader(w, r)
-	if err != nil {
-		http.Redirect(w, r, "/signin", http.StatusSeeOther)
-		return
-	}
-	type recipe struct {
-		Token
-		Adminsetting Adminsetting
-	}
-	rcp := recipe{}
-	rcp.Token = token
-	//mongoDB client 연결
-	client, err := mongo.NewClient(options.Client().ApplyURI(*flagMongoDBURI))
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	err = client.Connect(ctx)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	defer client.Disconnect(ctx)
-	err = client.Ping(ctx, readpref.Primary())
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	adminsetting, err := GetAdminSetting(client)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	rcp.Adminsetting = adminsetting
-	w.Header().Set("Content-Type", "text/html")
-	err = TEMPLATES.ExecuteTemplate(w, "cleanup-db-success", rcp)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
 	}
 }
